FD-2: the front door — Landing for guests, cockpit for authed, ?tab= is app intent
